Numerical matrix inversion works!

+++ src/geometry/Normal.cpp
@@ -8,7 +8,7 @@
 namespace RT {
 
     Normal::Normal() : x(0), y(0), z(0) {}
-    Normal::Normal(double x_init, double y_init, double z_init)
+    Normal::Normal(float x_init, float y_init, float z_init)
         : x(x_init), y(y_init), z(z_init) {}
 
     Normal Normal::operator+(const Normal& other) const
@@ -33,36 +33,36 @@ namespace RT {
         return *this;
     }
 
-    Normal Normal::operator*(const double scalar) const
+    Normal Normal::operator*(const float scalar) const
     {
         return Normal(x * scalar, y * scalar, z * scalar);
     }
 
-    Normal& Normal::operator*=(const double scalar)
+    Normal& Normal::operator*=(const float scalar)
     {
         x *= scalar; y *= scalar; z *= scalar;
         return *this;
     }
 
-    Normal Normal::operator/(const double scalar) const
+    Normal Normal::operator/(const float scalar) const
     {
-        const double inv = 1 / scalar;
+        const float inv = 1 / scalar;
         return Normal(x * inv, y * inv, z * inv);
     }
 
-    Normal& Normal::operator/=(const double scalar)
+    Normal& Normal::operator/=(const float scalar)
     {
-        const double inv = 1 / scalar;
+        const float inv = 1 / scalar;
         x *= inv; y *= inv; z *= inv;
         return *this;
     }
 
-    double Normal::length_squared() const
+    float Normal::length_squared() const
     {
         return x*x + y*y + z*z;
     }
 
-    double Normal::length() const
+    float Normal::length() const
     {
         return std::sqrt(length_squared());
     }
@@ -75,16 +75,16 @@ namespace RT {
 
     Normal normalize(const Normal& n)
     {
-        const double invLen = 1 / n.length();
+        const float invLen = 1 / n.length();
         return Normal(n.x * invLen, n.y * invLen, n.z * invLen);
     }
 
-    double dot(const Normal& n, const Vector3& v)
+    float dot(const Normal& n, const Vector3& v)
     {
             return n.x * v.x + n.y * v.y + n.z * v.z;
     }
 
-    Normal operator*(double s, const Normal& n)
+    Normal operator*(float s, const Normal& n)
     {
         return n * s;
     }
