+++ src/geometry/Point3.cpp
@@ -6,7 +6,7 @@
 namespace RT {
     
     Point3::Point3() : x(0), y(0), z(0) {}
-    Point3::Point3(double x_init, double y_init, double z_init)
+    Point3::Point3(float x_init, float y_init, float z_init)
         : x(x_init), y(y_init), z(z_init) {}
 
     Point3 Point3::operator+(const Vector3& v) const
@@ -37,12 +37,12 @@ namespace RT {
     }
 
 
-    double dist(const Point3& p1, const Point3& p2)
+    float dist(const Point3& p1, const Point3& p2)
     {
         return (p1 - p2).length();
     }
 
-    double squareDist(const Point3& p1, const Point3& p2)
+    float squareDist(const Point3& p1, const Point3& p2)
     {
         return (p1 - p2).length_squared();
     }
