+++ src/geometry/Vector3.cpp
@@ -2,10 +2,11 @@
 
 #include "Vector3.hpp"
 #include <cmath>
+
 namespace RT {
 
     Vector3::Vector3() : x(0), y(0), z(0) {}
-    Vector3::Vector3(double x_init, double y_init, double z_init)
+    Vector3::Vector3(float x_init, float y_init, float z_init)
         : x(x_init), y(y_init), z(z_init) {}
 
     Vector3 Vector3::operator+(const Vector3& other) const
@@ -30,26 +31,26 @@ namespace RT {
         return *this;
     }
 
-    Vector3 Vector3::operator*(const double scalar) const
+    Vector3 Vector3::operator*(const float scalar) const
     {
         return Vector3(x * scalar, y * scalar, z * scalar);
     }
 
-    Vector3& Vector3::operator*=(const double scalar)
+    Vector3& Vector3::operator*=(const float scalar)
     {
         x *= scalar; y *= scalar; z *= scalar;
         return *this;
     }
 
-    Vector3 Vector3::operator/(const double scalar) const
+    Vector3 Vector3::operator/(const float scalar) const
     {
-        const double inv = 1 / scalar;
+        const float inv = 1 / scalar;
         return Vector3(x * inv, y * inv, z * inv);
     }
 
-    Vector3& Vector3::operator/=(const double scalar)
+    Vector3& Vector3::operator/=(const float scalar)
     {
-        const double inv = 1 / scalar;
+        const float inv = 1 / scalar;
         x *= inv; y *= inv; z *= inv;
         return *this;
     }
@@ -59,12 +60,12 @@ namespace RT {
         return Vector3(-x, -y, -z);
     }
 
-    double Vector3::length_squared() const
+    float Vector3::length_squared() const
     {
         return x*x + y*y + z*z;
     }
 
-    double Vector3::length() const
+    float Vector3::length() const
     {
         return std::sqrt(length_squared());
     }
@@ -76,11 +77,11 @@ namespace RT {
 
     Vector3 normalize(const Vector3& vec)
     {
-        const double invLen = 1 / vec.length();
+        const float invLen = 1 / vec.length();
         return Vector3(vec.x * invLen, vec.y * invLen, vec.z * invLen);
     }
 
-    double dot(const Vector3& a, const Vector3& b)
+    float dot(const Vector3& a, const Vector3& b)
     {
         return a.x * b.x + a.y * b.y + a.z * b.z;
     }
@@ -92,7 +93,7 @@ namespace RT {
                        a.x * b.y - a.y * b.x);
     }
 
-    Vector3 operator*(double scalar, const Vector3& vec)
+    Vector3 operator*(float scalar, const Vector3& vec)
     {
         return vec * scalar;
     }
